Test Semantic Layer code. Add Levenshtein similarity metric. (#16)
In this PR I refactored the Semantic Layer code so that it is easier to
test, and I added a few unit and integration tests. I also added a
Levenshtein similarity metric to identify when the LLM or user misspells
a metric or dimension and suggest metrics and dimensions that are
similar to the misspelling.