Fix for multicell and empty Discovery API results (#23)
This PR adds support for MC environments. For MC, I decided to require
that the user add a separate env var for their MC account prefix. This
isn't ideal as more configuration is required, but I think it is more
reliable than trying to infer if the host is MC or not. This PR also
fixes errors being thrown when the Disco API returns empty results. I
don't have an MC account, so I wasn't able to test that, but I did test
that there were no regressions with my MT account.