Properly closing SL connection in the case of an exception (#53)
This PR ensures that the SL connection is closed in the event of an
exception. It is unclear to me why the connection was not being closed
properly before. I looked into the SL SDK code and read up on
`@contextmanager,` and the connection should have been closed, even if
there was an exception. However, apparently these connections weren't
being cleaned up properly, and this PR fixes it. I no longer see `Cannot
open session within session` errors.