Nest project under src (#54)
We are seeing problems with installing dbt-mcp:
```
❯ uvx --no-cache dbt-mcp
Traceback (most recent call last):
  File "/Users/devonfulcher/.local/share/uv/tools/dbt-mcp/bin/dbt-mcp", line 4, in <module>
    from dbt_mcp.main import main
ModuleNotFoundError: No module named 'dbt_mcp'
```
I think this may be due to the lack of an `src` directory, which seems
to be the standard way to layout a Python package. So this PR nests
`dbt_mcp` under `src`. As an example, the MCP Python SDK is laid out
this way: https://github.com/modelcontextprotocol/python-sdk. I tested
this by running the task commands that may be affected by this (e.g.
`task run`, `task test`, etc.). I will release a version after this is
merged to see if that fixes the problem.