dbt-mcp Installer 2000 (#66)
Installing the dbt-mcp server has never been easier!

All you need to run is:
```bash
/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/dbt-mcp/dbt-mcp/refs/heads/main/install.sh)"
```

And the installer will guide you through the installation and
configuration process.

Oh and it also acts as an updater, just run it again and it will update
the dbt-mcp server to the latest version, easy!

---------

Co-authored-by: BuildBot <buildbot@fishtownanalytics.com>
Co-authored-by: Devon Fulcher <24593113+DevonFulcher@users.noreply.github.com>