Usage tracking scaffolding (#116)
This PR sets up the codebase to emit usage tracking. This PR doesn't
emit events yet, but it should be straightforward to add that soon. This
is implemented by overriding the `call_tool` method and emitting events
after each tool call. This method is also used to catch exceptions, so I
was also able to remove some redundant try-except blocks.