Update Discovery config for simpler usage (#184)
This PR updates the Discovery config to make it a bit simpler to use.
This is relevant for our internal usage of this toolset, where the URL
is constructed differently from external usage. This update may also
have to be applied the SL toolset, but I just updated the Discovery
config for now.

I tested this with `task client` and utilized a Discovery tool with
that.