Catch every tool error and surface as string (#204)
This PR catches errors from tool calls and surfaces them as strings.
This should fix https://github.com/dbt-labs/dbt-mcp/issues/203. I was
able to reproduce the error and confirm that this solution fixes the
problem. It is unfortunate that this is necessary, but it seems Claude
specifically is having a hard time handling thrown exceptions with the
latest version of the MCP SDK.